+++ src/intranet3/scripts/twistd.py
@@ -1,24 +1,20 @@
-# -*- test-case-name: twisted.test.test_twistd -*-
-# Copyright (c) Twisted Matrix Laboratories.
-# See LICENSE for details.
-
-"""
-The Twisted Daemon: platform-independent interface.
-
-@author: Christopher Armstrong
-"""
-
 from twisted.application import app
 
-from _twistd_unix import ServerOptions, \
-    UnixApplicationRunner as _SomeApplicationRunner
+from twisted.scripts._twistd_unix import UnixApplicationRunner, ServerOptions, UnixAppLogger
+from twisted.web.server import Site
 
+from monkey_patches import _openLogFile, _getLogObserver
+
+def _monkey_patch_twistd_loggers():
+    UnixAppLogger._getLogObserver = _getLogObserver
+    Site._openLogFile = _openLogFile
 
 def runApp(config):
-    _SomeApplicationRunner(config).run()
+    UnixApplicationRunner(config).run()
 
 
 def run():
+    _monkey_patch_twistd_loggers()
     app.run(runApp, ServerOptions)