+++ src/intranet3/utils/mail.py
@@ -23,10 +23,10 @@ from twisted.mail.pop3client import POP3Client
 from twisted.internet.protocol import ClientFactory
 
 import transaction
-from intranet3.models import ApplicationConfig, TimeEntry, Project, Tracker, User, DBSession, TrackerCredentials
+from intranet3.models import ApplicationConfig, Project, Tracker, TrackerCredentials
 from intranet3.models.project import SelectorMapping
 from intranet3.log import DEBUG_LOG, WARN_LOG, EXCEPTION_LOG, INFO_LOG
-from intranet3.utils.smtp import ESMTP_XOUATH2_SenderFactory
+from intranet3.utils.timeentry import add_time
 
 try:
     from cStringIO import StringIO
@@ -392,38 +392,9 @@ class MailerPOP3Client(POP3Client):
         if data is None: # email should be ignored
             return
         user_id, date, bug_id, project_id, hours, subject = data
-
-        # try finding existing entry for this bug
-        session = DBSession()
-        bug_id = str(bug_id)
-        entry = TimeEntry.query.filter(TimeEntry.user_id==user_id)\
-                               .filter(TimeEntry.date==date.date())\
-                               .filter(TimeEntry.ticket_id==bug_id)\
-                               .filter(TimeEntry.project_id==project_id)\
-                               .first()
-        if not entry:
-            # create new entry
-            entry = TimeEntry(
-                user_id=user_id,
-                date=date.date(),
-                time=hours,
-                description=subject,
-                ticket_id=bug_id,
-                project_id = project_id,
-                modified_ts=date
-            )
-            session.add(entry)
-            LOG(u'Adding new entry')
-        else:
-            # update existing entry
-            if not entry.frozen:
-                entry.time += hours
-                entry.modified_ts = date # TODO: this might remove an already existing lateness
-                session.add(entry)
-                LOG(u'Updating existing entry')
-            else:
-                LOG(u'Omission of an existing entry because it is frozen')
+        add_time(user_id, date, bug_id, project_id, hours, subject)
         transaction.commit()
+
 
     def on_finish(self, results):
         """ When all messages have been retrieved """