+++ src/intranet3/views/presence.py
@@ -1,31 +1,31 @@
+# -*- coding: utf-8 -*-
 import datetime
 
+from babel.core import Locale
 from sqlalchemy import func
-from pyramid.httpexceptions import HTTPFound
 from pyramid.view import view_config
 
 from intranet3.utils.views import BaseView
 from intranet3.models import User, PresenceEntry
-from intranet3.helpers import previous_day, next_day
+from intranet3 import helpers as h
 from intranet3.utils import excuses
+
 
 day_start = datetime.time(0, 0, 0)
 day_end = datetime.time(23, 59, 59)
 hour_9 = datetime.time(9, 0, 0)
 
-
-@view_config(route_name='presence_today')
-class Today(BaseView):
-    def get(self):
-        today = datetime.datetime.now().date()
-        url = self.request.url_for('/presence/list', date=today.strftime('%d.%m.%Y'))
-        return HTTPFound(location=url)
+locale = Locale('en', 'US')
 
 
 @view_config(route_name='presence_list')
 class List(BaseView):
     def get(self):
-        date = datetime.datetime.strptime(self.request.GET.get('date'), '%d.%m.%Y')
+        date = self.request.GET.get('date')
+        if date:
+            date = datetime.datetime.strptime(date, '%d.%m.%Y')
+        else:
+            date = datetime.date.today()
         start_date = datetime.datetime.combine(date, day_start)
         end_date = datetime.datetime.combine(date, day_end)
         entries_p = self.session.query(User.id, User.name, func.min(PresenceEntry.ts), func.max(PresenceEntry.ts))\
@@ -48,16 +48,21 @@ class List(BaseView):
             entries_p=((user_id, user_name, start, stop, start.time() > hour_9) for (user_id, user_name, start, stop) in entries_p),
             entries_w=((user_id, user_name, start, stop, start.time() > hour_9) for (user_id, user_name, start, stop) in entries_w),
             date=date,
-            prev_date=previous_day(date),
-            next_date=next_day(date),
+            prev_date=h.previous_day(date),
+            next_date=h.next_day(date),
             excuses=excuses.presence(),
             justification=excuses.presence_status(date, self.request.user.id),
         )
 
+
 @view_config(route_name='presence_full')
 class Full(BaseView):
     def get(self):
-        date = datetime.datetime.strptime(self.request.GET.get('date'), '%d.%m.%Y')
+        date = self.request.GET.get('date')
+        if date:
+            date = datetime.datetime.strptime(date, '%d.%m.%Y')
+        else:
+            date = datetime.date.today()
 
         start_date = datetime.datetime.combine(date, day_start)
         end_date = datetime.datetime.combine(date, day_end)
@@ -69,5 +74,5 @@ class Full(BaseView):
         return dict(
             entries=entries,
             date=date,
-            prev_date=previous_day(date), next_date=next_day(date)
+            prev_date=h.previous_day(date), next_date=h.next_day(date)
         )